Выполнение задания 6 - Исправление логики обработки ошибок

+++ TaskJavaAdvanced06/Operation/Adding.java
@@ -3,7 +3,7 @@ package Operation;
 public class Adding implements Operation {
 
     @Override
-    public int result(int operandOne, int operandTwo) {
+    public double result(double operandOne, double operandTwo) {
         return operandOne + operandTwo;
     }
 }
