+++ TaskJavaAdvanced06/Operation/Multiplying.java
@@ -1,8 +1,9 @@
 package Operation;
 
 public class Multiplying implements Operation{
+
     @Override
-    public int result(int operandOne, int operandTwo) {
-        return operandOne*operandTwo;
+    public double result(double operandOne, double operandTwo) {
+        return operandOne * operandTwo;
     }
 }
